added validators for post call, authorized superuser and staff to edit/delete products and reformat

+++ Urn/views/products.py
@@ -5,11 +5,13 @@ from Urn.common import utils
 from collections import OrderedDict
 from Urn.models import Sku, Products, Businesses, ProductImages, BusinessUsers
 from django.views.decorators.csrf import csrf_exempt
+from Urn.schema_validators.products_validation import put_schema
 
 from Urn.schema_validators.sku_validation import schema
 from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseNotFound
 from Urn.common.utils import build_json, convert_uuid_string
-from Urn.decorators.validators import jwt_validate, check_authenticity, validate_schema, check_business_or_super
+from Urn.decorators.validators import jwt_validate, check_authenticity, validate_schema, check_business_or_super, \
+    validate_post_request_schema
 
 
 @csrf_exempt
@@ -126,8 +128,10 @@ def format_product_images(product_images, json=True):
 
 @csrf_exempt
 def process_products_request(request):
-    if request.method == 'POST':
+    if request.method == 'POST' and request.POST.get('_method', None) is None:
         return process_products_post(request)
+    elif request.method == 'POST' and request.POST.get('_method', None) == 'PUT':
+        return process_products_put(request)
     elif request.method == 'GET':
         return process_products_get(request)
     elif request.method == 'DELETE':
@@ -138,50 +142,55 @@ def process_products_request(request):
 
 @jwt_validate
 @check_business_or_super
+@validate_post_request_schema(schema)
 def process_products_post(request):
-    if request.method == 'POST' and request.POST.get('_method', None) is None:
-        request_data = json.loads(request.POST['product_json'])
-        files = request.FILES.getlist("product_images")
-        business_info = Businesses.objects.get(business_guid=request_data["business_guid"])
-        sku_info = Sku.objects.get(sku_guid=request_data["sku_guid"])
-        product = Products.objects.create(name=request_data["name"], description=request_data["description"],
-                                          price=request_data["price"], product_data=request_data["product_data"],
-                                          business_id=business_info.business_id, sku_id=sku_info.sku_id,
-                                          created_by=request.user.user_profile)
-        for file in files:
-            ProductImages.objects.create(product_id=product.product_id, image=file)
+    request_data = json.loads(request.POST['product_json'])
+    files = request.FILES.getlist("product_images")
+    business_info = Businesses.objects.get(business_guid=request_data["business_guid"])
+    sku_info = Sku.objects.get(sku_guid=request_data["sku_guid"])
+    product = Products.objects.create(name=request_data["name"], description=request_data["description"],
+                                      price=request_data["price"], product_data=request_data["product_data"],
+                                      business_id=business_info.business_id, sku_id=sku_info.sku_id,
+                                      created_by=request.user.user_profile)
+    for file in files:
+        ProductImages.objects.create(product_id=product.product_id, image=file)
 
-        return HttpResponse(status=201, content='Product added')
-
-    elif request.POST.get('_method', None) == "PUT":
-        return product_update_helper(request)
+    return HttpResponse(status=201, content='Product added')
 
 
-def product_update_helper(request):
+@jwt_validate
+@check_business_or_super
+@validate_post_request_schema(put_schema)
+def process_products_put(request):
+    user = request.user
+    user_info = request.user.user_profile
     request_data = json.loads(request.POST["product_json"])
     new_images = request.FILES.getlist("new_images")
     deleted_images = json.loads(request.POST["deleted_images"])
     product = Products.objects.filter(product_guid=request_data["product_guid"])
     if product.exists():
-        product_info = product.get()
-        sku_info = Sku.objects.get(sku_guid=request_data["sku_guid"])
-        request_data["updated_by"] = request.user.user_profile
-        request_data["sku_id"] = sku_info.sku_id
-        request_data["product_data"] = json.loads(request_data["product_data"])
-        del request_data["sku_guid"]
-        del request_data["business_guid"]
-        product.update(**request_data)
+        if BusinessUsers.objects.filter(business_id=product.get().business_id,
+                                        user_id=user_info.user_id).exists() or user.is_superuser or user.is_staff:
+            product_info = product.get()
+            sku_info = Sku.objects.get(sku_guid=request_data["sku_guid"])
+            request_data["updated_by"] = request.user.user_profile
+            request_data["sku_id"] = sku_info.sku_id
+            request_data["product_data"] = json.loads(request_data["product_data"])
+            del request_data["sku_guid"]
+            product.update(**request_data)
 
-        for image in new_images:
-            ProductImages.objects.create(product_id=product_info.product_id, image=image)
+            for image in new_images:
+                ProductImages.objects.create(product_id=product_info.product_id, image=image)
 
-        for image_guid in deleted_images:
-            image_info = ProductImages.objects.get(product_image_guid=image_guid)
-            if os.path.exists(image_info.image.url):
-                os.remove(image_info.image.url)
-            image_info.delete()
+            for image_guid in deleted_images:
+                image_info = ProductImages.objects.get(product_image_guid=image_guid)
+                if os.path.exists(image_info.image.url):
+                    os.remove(image_info.image.url)
+                image_info.delete()
 
-        return HttpResponse(status=202, content='Product updated')
+            return HttpResponse(status=202, content='Product updated')
+        else:
+            return HttpResponse(status=401, content='You are not authorized to edit this product')
     else:
         return HttpResponseBadRequest('No such product to update')
 
@@ -189,6 +198,8 @@ def product_update_helper(request):
 @jwt_validate
 @check_business_or_super
 def process_products_delete(request):
+    user = request.user
+    user_info = request.user.user_profile
     products_to_delete = json.loads(request.body.decode())["product_guid"]
     response = OrderedDict()
     response["success"] = list()
@@ -197,7 +208,7 @@ def process_products_delete(request):
         product_to_delete = Products.objects.filter(product_guid=each_product)
         if product_to_delete.exists():
             if BusinessUsers.objects.filter(business_id=product_to_delete.get().business_id,
-                                            user_id=request.user.user_profile.user_id).exists():
+                                            user_id=user_info.user_id).exists() or user.is_staff or user.is_superuser:
                 success_msg = "Product deleted : "
                 response["success"].append(success_msg + product_to_delete.get().name)
                 product_to_delete.delete()
